Extract Link signup controller from PayWithLinkViewController
